Storeを削除し、機能ごとに分割した
