+++ MainApp/EnableAzooKeyView/EnableAzooKeyView.swift
@@ -156,7 +156,7 @@ struct EnableAzooKeyView: View {
         .animation(.interactiveSpring(), value: step)
         .animation(.spring(), value: showDoneMessage)
         .onEnterForeground { _ in
-            if Store.checkKeyboardActivation() {
+            if SharedStore.checkKeyboardActivation() {
                 self.step = .setting
                 appStates.isKeyboardActivated = true
             }
