+++ MainApp/Theme/ThemeEditView.swift
@@ -251,7 +251,7 @@ struct ThemeEditView: CancelableEditor {
         case .themeShareView:
             ThemeShareView(theme: self.theme, shareImage: shareImage) {
                 presentationMode.wrappedValue.dismiss()
-                Store.shared.shouldTryRequestReview = true
+                RequestReviewManager.shared.shouldTryRequestReview = true
             }
             .navigationBarTitle(Text("完了"), displayMode: .inline)
             .navigationBarItems(leading: EmptyView(), trailing: EmptyView())
