+++ MainApp/Utils/MainAppDesign.swift
@@ -1,0 +1,11 @@
+//
+//  MainAppDesign.swift
+//  azooKey
+//
+//  Created by ensan on 2023/03/14.
+//  Copyright © 2023 ensan. All rights reserved.
+//
+
+enum MainAppDesign {
+    static let imageMaximumWidth: Double = 500
+}
