+++ MainApp/Utils/checkKeyboardActivation.swift
@@ -1,0 +1,19 @@
+//
+//  checkKeyboardActivation.swift
+//  azooKey
+//
+//  Created by ensan on 2023/03/14.
+//  Copyright © 2023 ensan. All rights reserved.
+//
+
+import Foundation
+
+extension SharedStore {
+    static func checkKeyboardActivation() -> Bool {
+        let bundleName = SharedStore.bundleName
+        guard let keyboards = UserDefaults.standard.dictionaryRepresentation()["AppleKeyboards"] as? [String] else {
+            return true
+        }
+        return keyboards.contains(bundleName)
+    }
+}
