+++ MainApp/Utils/closeKeyboard.swift
@@ -1,0 +1,15 @@
+//
+//  closeKeyboard.swift
+//  azooKey
+//
+//  Created by ensan on 2023/03/14.
+//  Copyright © 2023 ensan. All rights reserved.
+//
+
+import UIKit
+
+extension UIApplication {
+    func closeKeyboard() {
+        sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
+    }
+}
